Moves service creation to ServiceCore.from and adds some TODO notes
